backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pydantic==2.9.0
itsdangerous==2.2.0
httpx==0.27.2
pytest==8.3.3
python-dotenv==1.0.1
